+++ src/main/java/com/example/petShelter/service/ClientsService.java
@@ -10,7 +10,7 @@ import java.util.List;
 
 /**
  * The class consists of logic of the project, which has
- the methods  to work with "Clients" entity
+ * the methods  to work with "Clients" entity
  *
  * @author bornartem
  */
@@ -31,18 +31,24 @@ public class ClientsService {
      * @return create and save client in db
      */
     public Clients create(Clients clients) {
-        return clientsRepository.save(clients);
+        if (clientsRepository.existsById(clients.getId())) {
+            throw new RuntimeException();
+        } else {
+            return clientsRepository.save(clients);
+        }
     }
 
     /**
      * find client by id in db
      * called method of repository {@link JpaRepository#findById(Object)}
+     *
      * @param id identifier of client, can't be null
      * @return find client from db by id
      */
     public Clients read(long id) {
         return clientsRepository.findById(id).orElseThrow();
     }
+
     /**
      * update client in db
      * called method of repository {@link JpaRepository#save(Object)}
@@ -53,23 +59,34 @@ public class ClientsService {
     public Clients update(Clients clients) {
         return clientsRepository.save(clients);
     }
+
     /**
      * delete client by id in db
      * called method of repository {@link JpaRepository#findById(Object)}
+     *
      * @param id identifier of client, can't be null
      * @return delete client from db by id
      */
     public void delete(long id) {
-        clientsRepository.deleteById(id);
+        if (clientsRepository.existsById(id)) {
+            clientsRepository.deleteById(id);
+        } else {
+            throw new RuntimeException();
+        }
     }
 
     /**
      * get all clients from db
      * called method of repository {@link JpaRepository#findAll()}
+     *
      * @return all clients
      */
     public List<Clients> getAll() {
-        return clientsRepository.findAll();
+        if (clientsRepository.findAll().isEmpty()) {
+            throw new RuntimeException();
+        } else {
+            return clientsRepository.findAll();
+        }
     }
 
 }
